[I25-170] refactor : client, server 폴더 분리
